sleep after moving to get load info before asking for bootstrap token.  patch by Stu Hood; reviewed by jbellis for CASSANDRA-762
git-svn-id: https://svn.apache.org/repos/asf/incubator/cassandra/branches/cassandra-0.5@908267 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/cassandra/io/Streaming.java
@@ -51,7 +51,7 @@ public class Streaming
 {
     private static Logger logger = Logger.getLogger(Streaming.class);
     private static String TABLE_NAME = "STREAMING-TABLE-NAME";
-    public static final long RING_DELAY = 30 * 1000; // delay after which we assume ring has stablized
+    public static final int RING_DELAY = 30 * 1000; // delay after which we assume ring has stablized
 
     /**
      * Split out files for all tables on disk locally for each range and then stream them to the target endpoint.
